fix: byteSizeをu32にしライセンス一覧と検査を揃える
u64 は ts-rs が bigint へ写像するが、TauriのJSON IPCを通るとFrontendが
受け取るのは number であり乖離する。上限が10 MiBのため u32 とした。
ts-rs 追加分のライセンス一覧を再生成し、生成物がBiome対象外で検査から
漏れる問題に対して pre-commit へ再生成と差分検査を追加した。

+++ src-tauri/src/ipc/types.rs
@@ -67,7 +67,11 @@ pub struct FileContent {
     /// 判定した文字コード。BOMは `text` から除去済み。
     pub encoding: TextEncoding,
     /// 読み込んだバイト数。上限（10 MiB）の判定はRust側で行う。
-    pub byte_size: u64,
+    ///
+    /// TauriのJSON IPCは `serde_json` を通るため、Frontendが受け取るのはJavaScriptの
+    /// `number` である。`u64` は `ts-rs` が `bigint` へ写像し実値と乖離するため、
+    /// 上限が10 MiBであることを踏まえて `u32` とする。
+    pub byte_size: u32,
 }
 
 /// ファイル監視から届く変更の種別。
